+++ tests/Unit/Extension/Core/Rpc/StatusHandlerTest.php
@@ -39,6 +39,7 @@ class StatusHandlerTest extends HandlerTestCase
     public function testStatus()
     {
         $this->status->check()->willReturn([
+            'phpactor_version' => 'version one',
             'good' => [ 'i am good' ],
             'bad' => [ 'i am bad' ],
         ]);