+++ CHANGELOG.md
@@ -6,6 +6,11 @@ Changelog
 Bug fixes:
 
   - [Vim Plugin] Goto reference in a modified file causes warning #477.
+  - [Application] Improved status (show current version) #481
+
+## master
+
+  - [Docs] Added Development section
 
 ## 2018-05-20 0.5.0